merge it, homeboi

+++ routes/trail-routes.js
@@ -3,8 +3,8 @@ var db = require("../models");
 module.exports = function (app) {
 
     //require the express package and initialize it in the variable "app"
-    var express = require("express");
-    var app = express();
+    // var express = require("express");
+    // var app = express();
 
     //require the trails.js file through a direct pathway
     //var trails = require("../public/assets/js/trails.js");
@@ -13,7 +13,7 @@ module.exports = function (app) {
 
         db.Trail.findAll({
             where: {
-                trail_city: req.params.city
+                trailCity: req.params.city
             }
         }).then(function (dbTrail) {
             res.status(200).json(dbTrail);
@@ -24,27 +24,64 @@ module.exports = function (app) {
     });
 
     //route to display the "Add a new Trail" page
+
+
     app.post("/trails/add", function (req, res) {
-        console.log(req.body);
-        db.Trail.create({
-            // userName: req.body.userName,
-            trailName: req.body.trailName,
-            trailAddress: req.body.trailAddress,
-            trailCity: req.body.trailCity,
-            trailState: req.body.trailState,
-            trailLength: req.body.trailLength,
-            trailDifficulty: req.body.trailDifficulty
-        }).then(function (dbTrail) {
-            res.status(200).json({
-                trail: dbTrail,
-                success: true,
-                msg: "Trail successfully added."
+        if (req.isAuthenticated()) {
+            db.Trail.create({
+                userName: req.user.userName,
+                trailName: req.body.trailName,
+                trailAddress: req.body.trailAddress,
+                trailCity: req.body.trailCity,
+                trailState: req.body.trailState,
+                trailLength: req.body.trailLength,
+                trailDifficulty: req.body.trailDifficulty
+            }).then(function (dbTrail) {
+                res.status(200).json({
+                    trail: dbTrail,
+                    success: true,
+                    msg: "Trail successfully added."
+                });
+            }).catch(function (err) {
+                console.log(err);
             });
-        }).catch(function (err) {
-            console.log(err);
-        });
-        res.render("add-trail");
+
+            res.render("add-trail", user);
+        } else {
+            (function (dbTrail) {
+                res.status(401).json({
+                    msg: "You must be logged in to view this page."
+                });
+            });
+            res.redirect("/");
+        }
+
     });
+
+
+
+
+    // app.post("/trails/add", function (req, res) {
+    //     console.log(req.body);
+    //     db.Trail.create({
+    //         // userName: req.body.userName,
+    //         trailName: req.body.trailName,
+    //         trailAddress: req.body.trailAddress,
+    //         trailCity: req.body.trailCity,
+    //         trailState: req.body.trailState,
+    //         trailLength: req.body.trailLength,
+    //         trailDifficulty: req.body.trailDifficulty
+    //     }).then(function (dbTrail) {
+    //         res.status(200).json({
+    //             trail: dbTrail,
+    //             success: true,
+    //             msg: "Trail successfully added."
+    //         });
+    //     }).catch(function (err) {
+    //         console.log(err);
+    //     });
+    //     res.render("add-trail");
+    // });
 
     //route to display the "Update Trail" page
     app.put("/trails/update", function (req, res) {
@@ -63,21 +100,37 @@ module.exports = function (app) {
     });
 
     //route to display the "Delete Trail" warning
-    app.get("/trails/delete", function (req, res) {
-        db.Trail.destroy({
-            where: {
-                id: req.params.id
+    app.delete("/trails/delete", function (req, res) {
+        if (req.isAuthenticated()) {
+            if (req.user.accessLevel == "0") {
+                db.Trail.destroy({
+                    where: {
+                        id: req.params.id
+                    }
+                }).then(function (dbTrail) {
+                    res.json(dbTrail);
+                }).catch(function (err) {
+                    console.log(err);
+                });
+                res.render("delete-trail");
+            } else {
+                (function (dbTrail) {
+                    res.status(401).json({
+                        msg: "You must be an admin to delete this page."
+                    });
+                });
             }
-        }).then(function (dbTrail) {
-            res.json(dbTrail);
-        }).catch(function (err) {
-            console.log(err);
-        });
-        res.render("delete-trail");
+        } else {
+            (function (dbTrail) {
+                res.status(401).json({
+                    msg: "You must be logged in as admin to delete this page."
+                });
+            });
+            res.redirect("/");
+        }
+
+
+
+
     });
-
-
-
 }
-//export all functions with
-module.exports = app;
